Crate general_traits

This contains all the General-Traits for Tunneload that are used in a wide variety of places and that dont really have one single place to exist in other than this Crate

Traits

ConfigItem

Defines an interface for a simple Configuration that can easily be managed

DashboardEntity

The Bounds needed to register a new Entity on the Dashboard. This will provide all the Information that can then later be accessed and displayed on the Dashboard

DefaultConfig

A Trait that allows Configs to generate a Default-Configuration using a given Name

Handler

Defines a generic Handler that is responsible to handle incoming Connections, parse the Requests and route them to the right destinations as well as handling the responses

Receiver

This Trait specifies an interface that the Rest of the Codebase can use to read from an existing connection without needing to know about how this is actually done or through what acceptor this goes

Sender

This Trait specifies an interface that the Rest of the Codebase can use to send the Data back to the User, without needing to differentiate between having a normal Webserver serve the user or a connection from Tunneler