Trait general_traits::DashboardEntity

pub trait DashboardEntity {
    fn get_type(&self) -> &str;

    fn get_content(&self) -> Value;
}

The Bounds needed to register a new Entity on the Dashboard. This will provide all the Information that can then later be accessed and displayed on the Dashboard

Required methods

fn get_type(&self) -> &str

This should uniquely identify the Entity as it otherwise may lead to confusion when displaying the Data

This will be included in the form of a “type” Entry in the generated JSON object

fn get_content(&self) -> Value

This should return all the relevant Data for the given Entity

This will be included in the form of a “content” Entry in the generated JSON object