Trait general_traits::Receiver

pub trait Receiver: Send + Debug {
    #[must_use]
    fn read<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        buf: &'life1 mut [u8]
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn read_full<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        buf: &'life1 mut [u8]
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }
}

This Trait specifies an interface that the Rest of the Codebase can use to read from an existing connection without needing to know about how this is actually done or through what acceptor this goes

Required methods

#[must_use]
fn read<'life0, 'life1, 'async_trait>(
&'life0 mut self,
buf: &'life1 mut [u8]
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Reads from the Connection until there is either no more data left to read or until the provided Buffer is full

Returns: The number of bytes that were read from the connection and written into the provided Buffer

Provided methods

#[must_use]
fn read_full<'life0, 'life1, 'async_trait>(
&'life0 mut self,
buf: &'life1 mut [u8]
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Reads until the given Buffer is full or an error was encountered