Trait general_traits::Sender

pub trait Sender: Send + Debug {
    fn send<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        data: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn send_response<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self, 
        response: &'life1 Response<'life2>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait,
    { ... }
}

This Trait specifies an interface that the Rest of the Codebase can use to send the Data back to the User, without needing to differentiate between having a normal Webserver serve the user or a connection from Tunneler

Required methods

fn send<'life0, 'life1, 'async_trait>(
&'life0 mut self,
data: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Sends the given Piece of data

Provided methods

fn send_response<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
response: &'life1 Response<'life2>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Serializes and Sends the given Response using this Sender. This is just meant as a convience to not repeat the serialization and sending of responses in all sorts of situations